Added firebase and firestore initialization with generic converter types
