+++ src/firebase/config.ts
@@ -3,7 +3,7 @@ import { initializeApp } from 'firebase/app';
 import { initializeAuth, browserPopupRedirectResolver } from 'firebase/auth';
 import { getReactNativePersistence } from 'firebase/auth/react-native';
 // import { getAnalytics } from "firebase/analytics";
-import { getFirestore } from 'firebase/firestore';
+import { collection, CollectionReference, DocumentData, FirestoreDataConverter, getFirestore, QueryDocumentSnapshot, SnapshotOptions, WithFieldValue } from 'firebase/firestore';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
 // TODO: Add SDKs for Firebase products that you want to use
@@ -31,16 +31,43 @@ const firebaseConfig = {
 };
 
 // Initialize Firebase
-const app = initializeApp(firebaseConfig);
+export const app = initializeApp(firebaseConfig);
+
 /* 
   to get rid of async storage incompatibility with expo, initialize auth with the following method
   https://github.com/firebase/firebase-js-sdk/issues/1847#issuecomment-915634151
 */
-const auth = initializeAuth(app, {
+export const auth = initializeAuth(app, {
     persistence: [getReactNativePersistence(AsyncStorage)],
     popupRedirectResolver: browserPopupRedirectResolver,
 });
-// const analytics = getAnalytics(app);
-const db = getFirestore(app);
 
-export { app, auth, db };
+/*
+    Initialize firestore and define typed helping collection function
+        Reference: https://plainenglish.io/blog/using-firestore-with-typescript-in-the-v9-sdk-cf36851bb099
+        https://medium.com/swlh/using-firestore-with-typescript-65bd2a602945
+*/
+export const db = getFirestore(app);
+
+// Generic data type converter from firestore
+const genericConverter = <T>() => ({
+    toFirestore: (inputData: T) => inputData,
+    fromFirestore: (
+      snapshot: QueryDocumentSnapshot,
+      options: SnapshotOptions
+    ): T => snapshot.data() as T,
+});
+
+// create a collection function, using typecasting and the withConverter function to get typed data back from firestore
+const createCollection = <T = DocumentData>(collectionName: string) => {
+    const converter = genericConverter<T>() as FirestoreDataConverter<T>;
+    return collection(db, collectionName).withConverter<T>(converter);
+};
+
+/*
+    Define the collections
+*/
+export const privateUserCollection = createCollection('private-user-data');
+export const publicUserCollection = createCollection('public-user-data');
+
+
